Turn process-extras-tests into a test-suite
This may be exhibiting cabal bug #1810, if you run with --show-details=streaming
you will see it go through all the tests but seemingly hang to exit.  I left it
running overnight and it exited successfully, but I didn't note how long it took.
It is likely that this only affects the particular version of Cabal that i am
using - 1.20 modified for ghcjs.